requirements/base.txt
django==2.1.1
mistune==0.8.3
pygments==2.2.0
django-compressor==2.2
django-pagedown==1.0.6
django-uuslug==1.1.8
